+++ shared/Beatmaps/NetworkBeatmapLevel.hpp
@@ -10,8 +10,8 @@ DECLARE_CLASS_CUSTOM(MultiplayerCore::Beatmaps, NetworkBeatmapLevel, Abstraction
     DECLARE_CTOR(ctor_1, Packets::MpBeatmapPacket*);
 
     public:
-        const std::unordered_map<uint8_t, std::list<std::string>>& get_requirements() { return _packet->requirements; };
-        const std::unordered_map<uint8_t, MultiplayerCore::Beatmaps::Abstractions::DifficultyColors>& get_mapColors() { return _packet->mapColors; };
+        const std::unordered_map<uint32_t, std::list<std::string>>& get_requirements() { return _packet->requirements; };
+        const std::unordered_map<uint32_t, MultiplayerCore::Beatmaps::Abstractions::DifficultyColors>& get_mapColors() { return _packet->mapColors; };
         const std::vector<const MultiplayerCore::Utils::ExtraSongData::Contributor>& get_contributors() { return _packet->contributors; };
 )
