+++ shared/Beatmaps/Packets/MpBeatmapPacket.hpp
@@ -27,7 +27,7 @@ DECLARE_CLASS_CUSTOM(MultiplayerCore::Beatmaps::Packets, MpBeatmapPacket, Multip
     DECLARE_CTOR(New);
     public:
         static MpBeatmapPacket* New_1(GlobalNamespace::BeatmapLevel* beatmap, GlobalNamespace::BeatmapKey beatmapKey);
-        std::unordered_map<uint8_t, std::list<std::string>> requirements;
-        std::unordered_map<uint8_t, MultiplayerCore::Beatmaps::Abstractions::DifficultyColors> mapColors;
+        std::unordered_map<uint32_t, std::list<std::string>> requirements;
+        std::unordered_map<uint32_t, MultiplayerCore::Beatmaps::Abstractions::DifficultyColors> mapColors;
         std::vector<const MultiplayerCore::Utils::ExtraSongData::Contributor> contributors;
 )
